refactor: to subdir /token-aggregators AND benchmarks display protocol name
